basic tetris files
This is the code I have input working with Ania Kubow's JS Tetris tutorial. I got to the point were it would create a tetramino and it would start to drop. It stops at the bottom. Then it is supposed to draw a new tetramino but it does not. I tried to move on after it didn't work but I still did not get it to work.